+++ b4n/kube/resources/resources_list.rs
@@ -184,6 +184,7 @@ impl Table for ResourcesList {
             fn get_highlighted_item_index(&self) -> Option<usize>;
             fn get_highlighted_item_name(&self) -> Option<&str>;
             fn get_highlighted_item_uid(&self) -> Option<&str>;
+            fn get_highlighted_item_line_no(&self) -> Option<u16>;
             fn highlight_item_by_name(&mut self, name: &str) -> bool;
             fn highlight_item_by_name_start(&mut self, text: &str) -> bool;
             fn highlight_item_by_uid(&mut self, uid: &str) -> bool;