+++ b4n/ui/views/resources/view.rs
@@ -287,6 +287,20 @@ impl ResourcesView {
             return ResponseEvent::Handled;
         }
 
+        if self.app_data.has_binding(event, KeyCommand::MouseMenuOpen)
+            && let Some(line_no) = self.table.list.table.get_highlighted_item_line_no()
+            && let Some(name) = self.table.list.table.get_highlighted_item_name()
+        {
+            let view = self.table.list.table.table.header.get_cached_view();
+            let width = self.table.list.table.table.header.get_cached_width().unwrap_or_default();
+            let (namespace_width, _, _) = self.table.list.table.table.header.get_widths(view, width);
+            self.show_mouse_menu(
+                u16::try_from(namespace_width + name.chars().count() + 6).unwrap_or_default(),
+                line_no.saturating_add(self.table.list.area.y),
+            );
+            return ResponseEvent::Handled;
+        }
+
         if self.app_data.has_binding(event, KeyCommand::YamlCreate) {
             self.show_create_resource_palette();
             return ResponseEvent::Handled;